requirements.txt
flask
openai
langchain-community
langchain-text-splitters
sentence-transformers
rank-bm25
faiss-cpu
numpy
requests
docling
pypdf
docx2txt
wordsegment